refactor: Member 도메인 클래스명 변경
회원가입/로그인 등의 기능을 AuthController, AuthService 에서 처리하도록 변경
