+++ src/main/java/sws/songpin/domain/member/controller/AuthController.java
@@ -13,24 +13,25 @@ import org.springframework.web.bind.annotation.RestController;
 import sws.songpin.domain.member.dto.request.LoginRequestDto;
 import sws.songpin.domain.member.dto.request.SignUpRequestDto;
 import sws.songpin.domain.member.dto.response.LoginResponseDto;
+import sws.songpin.domain.member.service.AuthService;
 import sws.songpin.domain.member.service.MemberService;
 
 @RestController
 @RequiredArgsConstructor
-public class LoginController {
-    private final MemberService memberService;
+public class AuthController {
+    private final AuthService authService;
 
     @Operation(summary = "회원가입", description = "회원 가입을 통해 유저 생성")
     @PostMapping("/signup")
     public ResponseEntity<Object> signUp(@Valid @RequestBody SignUpRequestDto requestDto){
-        memberService.signUp(requestDto);
+        authService.signUp(requestDto);
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @Operation(summary = "로그인", description = "로그인 결과를 반환합니다.")
     @PostMapping("/login")
     public ResponseEntity<LoginResponseDto> login(@Valid @RequestBody LoginRequestDto requestDto){
-        LoginResponseDto responseDto = memberService.login(requestDto);
+        LoginResponseDto responseDto = authService.login(requestDto);
         return ResponseEntity.ok(responseDto);
     }
